Compile fix for plate2plate
Only effects 32bit systems

+++ src/vw/Plate/plate2plate.cc
@@ -111,7 +111,7 @@ struct ToastDem : public FilterBase<ToastDem> {
   struct DemWriter : public ToastDemWriter {
     PlateFile& platefile;
     DemWriter(PlateFile& output) : platefile(output) { }
-    inline void operator()(const boost::shared_array<uint8> data, uint64 data_size,
+    inline void operator()(const boost::shared_array<uint8> data, size_t data_size,
                            int32 dem_col, int32 dem_row,
                            int32 dem_level, int32 output_transaction_id) const {
       platefile.write_update(data, data_size, dem_col, dem_row,
